+++ src/reflection/collection/ReflectionCollection.php
@@ -17,7 +17,7 @@ use PhpCollection\Map;
 use cloak\collection\PairStackable;
 use cloak\reflection\Reflection;
 use cloak\reflection\ResultConvertible;
-use cloak\CollectionInterface;
+use cloak\Collection;
 use cloak\result\LineResultSelectable;
 use cloak\result\collection\CoverageResultCollection;
 use \Closure;
@@ -29,7 +29,7 @@ use \ArrayIterator;
  * Class ReflectionCollection
  * @package cloak\reflection\collection
  */
-class ReflectionCollection implements CollectionInterface, ResultCollectionConvertible
+class ReflectionCollection implements Collection, ResultCollectionConvertible
 {
 
     use PairStackable;